+++ TraceBook/src/de/fu/tracebook/gui/activity/AddPointActivity.java
@@ -313,7 +313,7 @@ public class AddPointActivity extends ListActivity {
                 && object instanceof IDataNode
                 && ((IDataNode) object).getDataPointsList() == null) {
             LogIt.d("POI is empty, will not keep it");
-            Helper.currentTrack().deleteNode((int) object.getId());
+            Helper.currentTrack().deleteNode(object.getId());
             (new GpsMessage(this)).sendDiscardIntent();
         }
